+++ CalculateBonus/src/models/salaried/Salaried.java
@@ -4,7 +4,7 @@ import domains.Person;
 
 public class Salaried extends Person {
 
-    public Salaried(String name, String lastName, int age, String dni, String address, double salary) {
+    public Salaried(String name, String lastName, int age, String dni, String address, float salary) {
         super(name, lastName, age, dni, address, salary);
     }
