+++ CalculateBonus/src/models/staff/AdministrativeStaff.java
@@ -4,7 +4,7 @@ import models.salaried.Salaried;
 
 public class AdministrativeStaff extends Salaried {
 
-    public AdministrativeStaff(String name, String lastName, int age, String dni, String address, double salary) {
+    public AdministrativeStaff(String name, String lastName, int age, String dni, String address, float salary) {
         super(name, lastName, age, dni, address, salary);
     }
