+++ CalculateBonus/src/models/teachers/FullTeacher.java
@@ -4,7 +4,7 @@ public class FullTeacher extends Teacher {
 
     private int yearsOfExperience;
 
-    public FullTeacher(String name, String lastName, int age, String dni, String address, double salary) {
+    public FullTeacher(String name, String lastName, int age, String dni, String address, float salary) {
         super(name, lastName, age, dni, address, salary);
     }
