+++ CalculateBonus/src/models/teachers/InterimTeacher.java
@@ -2,7 +2,7 @@ package models.teachers;
 
 public class InterimTeacher extends Teacher {
 
-    public InterimTeacher(String name, String lastName, int age, String dni, String address, double salary) {
+    public InterimTeacher(String name, String lastName, int age, String dni, String address, float salary) {
         super(name, lastName, age, dni, address, salary);
     }
 }
